_update_scope belongs into the scope not the client
