+++ sentry_sdk/scope.py
@@ -81,6 +81,24 @@ def _disable_capture(fn):
     return wrapper  # type: ignore
 
 
+def _update_scope(base, scope_change, scope_kwargs):
+    # type: (Scope, Optional[Any], Dict[str, Any]) -> Scope
+    if scope_change and scope_kwargs:
+        raise TypeError("cannot provide scope and kwargs")
+    if scope_change is not None:
+        final_scope = copy(base)
+        if callable(scope_change):
+            scope_change(final_scope)
+        else:
+            final_scope.update_from_scope(scope_change)
+    elif scope_kwargs:
+        final_scope = copy(base)
+        final_scope.update_from_kwargs(**scope_kwargs)
+    else:
+        final_scope = base
+    return final_scope
+
+
 class Scope(object):
     """The scope holds extra information that should be sent with all
     events that belong to it.
